feat : service 작성 및 반환 타입 수정

+++ src/main/java/com/planti/domain/community/dto/response/CommentDto.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Builder
 public class CommentDto {
-    private Integer commentId;
+    private long commentId;
     private Long userId;
     private String nickname;
     private String content;
